refactor(#3721): keep every touched file at or under its campsite baseline
The file-size ratchet (correctly) FAILed: my comment blocks grew four
over-threshold files. Nothing is waived — the prose moves to where the policy
already lives and one function moves to its natural home.

  big_promoted.rs           800 -> 758  (was 823)
  data_access/mod.rs        982 -> 982  (was 987)
  block_emit_windowed.rs   1012 -> 1012 (was 1045)
  row_data.rs               917 -> 917  (was 952)
  bti.rs                    827 -> 803

* The cell-flags authority, the phantom-partition measurement and the
  per-caller fallback rationale move into `column_decode_error`'s module docs
  (310 lines), leaving pointers at the three call sites. One place states the
  policy, which is where a reader looks for it anyway.
* `big_decode_clustering_window` moves from `big_promoted.rs` into
  `clustering_seek_decode.rs`, whose sole caller it now is and which is
  literally named for it. `decompress_partition_window` becomes `pub(super)`.
* `clustering_seek_decode` is declared from `bti.rs` (under threshold and
  shrinking) rather than `data_access/mod.rs` (982 and over), and lives at
  `data_access/bti/` to match — it serves `bti.rs`'s seek and nothing else.

issue_3721 17/17, issue_1184 6/6, issue_993 5/5, point_vs_full 8/8; clippy
clean at default, tombstones and cli-helpers.
